+++ packages/terra-form-input/CHANGELOG.md
@@ -3,11 +3,22 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Updated theme variables
+
+1.32.0 - (January 2, 2019)
+------------------
 ### Fixed
 * Dev site tests pages adjusted to uniformly use caretColor
+
 ### Added
-* Pattern prop added to Input component 
+* Pattern prop added to Input component
 * maxWidth prop added to pass through to terra-form-field
+
+### Changed
+* Minor dependency version bump
+* Updated site examples to be more accessible.
+* Note added regarding input types to pattern prop
 
 1.31.1 - (November 20, 2018)
 ------------------